move port test case to passed
